+++ lib/views/diary/diary_post_view.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:nutripic/components/custom_app_bar.dart';
+import 'package:nutripic/components/common/custom_app_bar.dart';
 import 'package:nutripic/view_models/diary/diary_post_view_model.dart';
 import 'package:provider/provider.dart';
