+++ lib/views/login/signup_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:nutripic/components/custom_app_bar.dart';
+import 'package:nutripic/components/common/custom_app_bar.dart';
 import 'package:nutripic/view_models/login/signup_view_model.dart';
 import 'package:provider/provider.dart';
