+++ lib/views/user_info/user_edit_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:nutripic/components/button_tile.dart';
-import 'package:nutripic/components/custom_app_bar.dart';
-import 'package:nutripic/components/custom_scaffold.dart';
+import 'package:nutripic/components/common/custom_app_bar.dart';
+import 'package:nutripic/components/common/custom_scaffold.dart';
 import 'package:nutripic/components/custom_text_field.dart';
 import 'package:nutripic/components/user_info/profile_image.dart';
 import 'package:nutripic/main.dart';
